+++ exp/orderbook/pools.go
@@ -2,8 +2,8 @@ package orderbook
 
 import (
 	"math"
-	"math/big"
 
+	"github.com/holiman/uint256"
 	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/xdr"
 )
@@ -91,8 +91,8 @@ func makeTrade(
 //
 // It returns false if the calculation overflows.
 func calculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int32) (xdr.Int64, bool) {
-	X, Y := big.NewInt(int64(reserveA)), big.NewInt(int64(reserveB))
-	F, x := big.NewInt(int64(feeBips)), big.NewInt(int64(received))
+	X, Y := uint256.NewInt(uint64(reserveA)), uint256.NewInt(uint64(reserveB))
+	F, x := uint256.NewInt(uint64(feeBips)), uint256.NewInt(uint64(received))
 
 	// would this deposit overflow the reserve?
 	if received > math.MaxInt64-reserveA {
@@ -100,12 +100,12 @@ func calculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int
 	}
 
 	// We do all of the math in bips, so it's all upscaled by this value.
-	maxBips := big.NewInt(10000)
-	f := new(big.Int).Sub(maxBips, F) // upscaled 1 - F
+	maxBips := uint256.NewInt(10000)
+	f := new(uint256.Int).Sub(maxBips, F) // upscaled 1 - F
 
 	// right half: X + (1 - F)x
-	denom := X.Mul(X, maxBips).Add(X, new(big.Int).Mul(x, f))
-	if denom.Cmp(big.NewInt(0)) == 0 { // avoid div-by-zero panic
+	denom := X.Mul(X, maxBips).Add(X, new(uint256.Int).Mul(x, f))
+	if denom.IsZero() { // avoid div-by-zero panic
 		return 0, false
 	}
 
@@ -115,8 +115,8 @@ func calculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int
 	// divide & check overflow
 	result := numer.Div(numer, denom)
 
-	i := xdr.Int64(result.Int64())
-	return i, result.IsInt64() && i > 0
+	val := xdr.Int64(result.Uint64())
+	return val, result.IsUint64() && val >= 0
 }
 
 // calculatePoolExpectation determines how much of `reserveA` you would need to
@@ -128,8 +128,8 @@ func calculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int
 func calculatePoolExpectation(
 	reserveA, reserveB, disbursed xdr.Int64, feeBips xdr.Int32,
 ) (xdr.Int64, bool) {
-	X, Y := big.NewInt(int64(reserveA)), big.NewInt(int64(reserveB))
-	F, y := big.NewInt(int64(feeBips)), big.NewInt(int64(disbursed))
+	X, Y := uint256.NewInt(uint64(reserveA)), uint256.NewInt(uint64(reserveB))
+	F, y := uint256.NewInt(uint64(feeBips)), uint256.NewInt(uint64(disbursed))
 
 	// sanity check: disbursing shouldn't underflow the reserve
 	if disbursed >= reserveB {
@@ -137,25 +137,27 @@ func calculatePoolExpectation(
 	}
 
 	// We do all of the math in bips, so it's all upscaled by this value.
-	maxBips := big.NewInt(10000)
-	f := new(big.Int).Sub(maxBips, F) // upscaled 1 - F
+	maxBips := uint256.NewInt(10000)
+	f := new(uint256.Int).Sub(maxBips, F) // upscaled 1 - F
 
-	denom := Y.Sub(Y, y).Mul(Y, f)     // right half: (Y - y)(1 - F)
-	if denom.Cmp(big.NewInt(0)) == 0 { // avoid div-by-zero panic
+	denom := Y.Sub(Y, y).Mul(Y, f) // right half: (Y - y)(1 - F)
+	if denom.IsZero() {            // avoid div-by-zero panic
 		return 0, false
 	}
 
 	numer := X.Mul(X, y).Mul(X, maxBips) // left half: Xy
 
-	result, rem := new(big.Int), new(big.Int)
-	result.DivMod(numer, denom, rem)
+	result, rem := new(uint256.Int), new(uint256.Int)
+	result.Div(numer, denom)
+	rem.Mod(numer, denom)
 
 	// hacky way to ceil(): if there's a remainder, add 1
-	if rem.Cmp(big.NewInt(0)) > 0 {
-		result.Add(result, big.NewInt(1))
+	if !rem.IsZero() {
+		result.AddUint64(result, 1)
 	}
 
-	return xdr.Int64(result.Int64()), result.IsInt64()
+	val := xdr.Int64(result.Uint64())
+	return val, result.IsUint64() && val >= 0
 }
 
 // getOtherAsset returns the other asset in the liquidity pool. Note that