+++ services/horizon/internal/integration/db_test.go
@@ -3,12 +3,16 @@ package integration
 import (
 	"context"
 	"fmt"
-	horizon "github.com/stellar/go/services/horizon/internal"
-	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"path/filepath"
 	"strconv"
 	"testing"
 	"time"
+
+	"github.com/stellar/go/clients/horizonclient"
+	"github.com/stellar/go/keypair"
+	horizon "github.com/stellar/go/services/horizon/internal"
+	"github.com/stellar/go/services/horizon/internal/db2/history"
+	"github.com/stellar/go/xdr"
 
 	"github.com/stretchr/testify/assert"
 
@@ -21,22 +25,123 @@ import (
 	"github.com/stellar/go/txnbuild"
 )
 
-func initializeDBIntegrationTest(t *testing.T) (itest *integration.Test, reachedLedger int32) {
-	itest = integration.NewTest(t, protocol15Config)
+func generateLiquidityPoolOps(itest *integration.Test, tt *assert.Assertions) (lastLedger int32) {
+
 	master := itest.Master()
+	keys, accounts := itest.CreateAccounts(2, "1000")
+	shareKeys, shareAccount := keys[0], accounts[0]
+	tradeKeys, tradeAccount := keys[1], accounts[1]
+
+	itest.MustSubmitMultiSigOperations(shareAccount, []*keypair.Full{shareKeys, master},
+		&txnbuild.ChangeTrust{
+			Line: txnbuild.ChangeTrustAssetWrapper{
+				Asset: txnbuild.CreditAsset{
+					Code:   "USD",
+					Issuer: master.Address(),
+				},
+			},
+			Limit: txnbuild.MaxTrustlineLimit,
+		},
+		&txnbuild.ChangeTrust{
+			Line: txnbuild.LiquidityPoolShareChangeTrustAsset{
+				LiquidityPoolParameters: txnbuild.LiquidityPoolParameters{
+					AssetA: txnbuild.NativeAsset{},
+					AssetB: txnbuild.CreditAsset{
+						Code:   "USD",
+						Issuer: master.Address(),
+					},
+					Fee: 30,
+				},
+			},
+			Limit: txnbuild.MaxTrustlineLimit,
+		},
+		&txnbuild.Payment{
+			SourceAccount: master.Address(),
+			Destination:   shareAccount.GetAccountID(),
+			Asset: txnbuild.CreditAsset{
+				Code:   "USD",
+				Issuer: master.Address(),
+			},
+			Amount: "1000",
+		},
+	)
+
+	poolID, err := xdr.NewPoolId(
+		xdr.MustNewNativeAsset(),
+		xdr.MustNewCreditAsset("USD", master.Address()),
+		30,
+	)
+	tt.NoError(err)
+	poolIDHexString := xdr.Hash(poolID).HexString()
+
+	itest.MustSubmitOperations(shareAccount, shareKeys,
+		&txnbuild.LiquidityPoolDeposit{
+			LiquidityPoolID: [32]byte(poolID),
+			MaxAmountA:      "400",
+			MaxAmountB:      "777",
+			MinPrice:        "0.5",
+			MaxPrice:        "2",
+		},
+	)
+
+	itest.MustSubmitOperations(tradeAccount, tradeKeys,
+		&txnbuild.ChangeTrust{
+			Line: txnbuild.ChangeTrustAssetWrapper{
+				Asset: txnbuild.CreditAsset{
+					Code:   "USD",
+					Issuer: master.Address(),
+				},
+			},
+			Limit: txnbuild.MaxTrustlineLimit,
+		},
+		&txnbuild.PathPaymentStrictReceive{
+			SendAsset: txnbuild.NativeAsset{},
+			DestAsset: txnbuild.CreditAsset{
+				Code:   "USD",
+				Issuer: master.Address(),
+			},
+			SendMax:     "1000",
+			DestAmount:  "2",
+			Destination: tradeKeys.Address(),
+		},
+	)
+
+	pool, err := itest.Client().LiquidityPoolDetail(horizonclient.LiquidityPoolRequest{
+		LiquidityPoolID: poolIDHexString,
+	})
+	tt.NoError(err)
+
+	txResp := itest.MustSubmitOperations(shareAccount, shareKeys,
+		&txnbuild.LiquidityPoolWithdraw{
+			LiquidityPoolID: [32]byte(poolID),
+			Amount:          pool.TotalShares,
+			MinAmountA:      "10",
+			MinAmountB:      "20",
+		},
+	)
+
+	return txResp.Ledger
+}
+
+func generatePaymentOps(itest *integration.Test, tt *assert.Assertions) (lastLedger int32) {
+	txResp := itest.MustSubmitOperations(itest.MasterAccount(), itest.Master(),
+		&txnbuild.Payment{
+			Destination: itest.Master().Address(),
+			Amount:      "10",
+			Asset:       txnbuild.NativeAsset{},
+		},
+	)
+
+	return txResp.Ledger
+}
+
+func initializeDBIntegrationTest(t *testing.T) (itest *integration.Test, reachedLedger int32) {
+	config := integration.Config{ProtocolVersion: 18}
+	itest = integration.NewTest(t, config)
 	tt := assert.New(t)
 
-	// Initialize the database with some ledgers including some transactions we submit
-	op := txnbuild.Payment{
-		Destination: master.Address(),
-		Amount:      "10",
-		Asset:       txnbuild.NativeAsset{},
-	}
-	// TODO: should we enforce certain number of ledgers to be ingested?
-	for i := 0; i < 8; i++ {
-		txResp := itest.MustSubmitOperations(itest.MasterAccount(), master, &op)
-		reachedLedger = txResp.Ledger
-	}
+	generatePaymentOps(itest, tt)
+	reachedLedger = generateLiquidityPoolOps(itest, tt)
 
 	root, err := itest.Client().Root()
 	tt.NoError(err)
@@ -59,9 +164,23 @@ func TestReingestDB(t *testing.T) {
 	horizonConfig.DatabaseURL = freshHorizonPostgresURL
 	// Initialize the DB schema
 	dbConn, err := db.Open("postgres", freshHorizonPostgresURL)
+	tt.NoError(err)
 	defer dbConn.Close()
 	_, err = schema.Migrate(dbConn.DB.DB, schema.MigrateUp, 0)
 	tt.NoError(err)
+
+	t.Run("validate parallel range", func(t *testing.T) {
+		horizoncmd.RootCmd.SetArgs(command(horizonConfig,
+			"db",
+			"reingest",
+			"range",
+			"--parallel-workers=2",
+			"10",
+			"2",
+		))
+
+		assert.EqualError(t, horizoncmd.RootCmd.Execute(), "Invalid range: {10 2} from > to")
+	})
 
 	// cap reachedLedger to the nearest checkpoint ledger because reingest range cannot ingest past the most
 	// recent checkpoint ledger when using captive core
@@ -99,12 +218,13 @@ func TestReingestDB(t *testing.T) {
 	horizoncmd.RootCmd.SetArgs(command(horizonConfig, "db",
 		"reingest",
 		"range",
+		"--parallel-workers=1",
 		"1",
 		fmt.Sprintf("%d", toLedger),
 	))
 
-	// Reingest into the DB
 	tt.NoError(horizoncmd.RootCmd.Execute())
+	tt.NoError(horizoncmd.RootCmd.Execute(), "Repeat the same reingest range against db, should not have errors.")
 }
 
 func command(horizonConfig horizon.Config, args ...string) []string {
@@ -157,6 +277,18 @@ func TestFillGaps(t *testing.T) {
 	})
 	tt.NoError(err)
 
+	t.Run("validate parallel range", func(t *testing.T) {
+		horizoncmd.RootCmd.SetArgs(command(horizonConfig,
+			"db",
+			"fill-gaps",
+			"--parallel-workers=2",
+			"10",
+			"2",
+		))
+
+		assert.EqualError(t, horizoncmd.RootCmd.Execute(), "Invalid range: {10 2} from > to")
+	})
+
 	// make sure a full checkpoint has elapsed otherwise there will be nothing to reingest
 	var latestCheckpoint uint32
 	publishedFirstCheckpoint := func() bool {
@@ -186,7 +318,7 @@ func TestFillGaps(t *testing.T) {
 		filepath.Dir(horizonConfig.CaptiveCoreConfigPath),
 		"captive-core-reingest-range-integration-tests.cfg",
 	)
-	horizoncmd.RootCmd.SetArgs(command(horizonConfig, "db", "fill-gaps"))
+	horizoncmd.RootCmd.SetArgs(command(horizonConfig, "db", "fill-gaps", "--parallel-workers=1"))
 	tt.NoError(horizoncmd.RootCmd.Execute())
 
 	tt.NoError(historyQ.LatestLedger(context.Background(), &latestLedger))